requirements.txt
Django==3.1.7
Pillow==8.1.2
django-fontawesome==1.0
python-slugify==4.0.1